Removed implementations and refactor some codes
